+++ gdsfactory/components/cdc.py
@@ -78,7 +78,7 @@ def cdc(
         direction=direction,
     )
 
-    return gf.read.picwriter(c)
+    return gf.read.from_picwriter(c)
 
 
 if __name__ == "__main__":