+++ gdsfactory/components/dbr2.py
@@ -92,7 +92,7 @@ def dbr2(
         direction=direction,
     )
 
-    return gf.read.picwriter(c)
+    return gf.read.from_picwriter(c)
 
 
 if __name__ == "__main__":