+++ gdsfactory/read/from_picwriter.py
@@ -53,7 +53,9 @@ def direction_to_degree(direction: str) -> float:
     return direction_to_orientation[direction]
 
 
-def picwriter(picwriter_object: pt.Component, port_layer: Layer = (1, 0)) -> Component:
+def from_picwriter(
+    picwriter_object: pt.Component, port_layer: Layer = (1, 0)
+) -> Component:
     """Convert a Picwriter into a Gdsfactory Component."""
     po = picwriter_object
     c = gf.Component(name=po.name_prefix)
@@ -108,6 +110,6 @@ if __name__ == "__main__":
 
     # gc = pc.GratingCoupler(wgt, port=(10, 20), direction=np.pi * 7 / 8)
     gc = pc.GratingCoupler(wgt, port=(10, 20), direction=0.0)
-    gcc = picwriter(gc)
+    gcc = from_picwriter(gc)
 
     gf.show(gcc)